Added cmus theme.
The theme isn't complete yet, there're still things I would like to do
to it to make it better. I think that cmus doesn't have good themeing
notes and doesn't specify all of the options. The box containing the
album name in the track listing on the right-hand side's background
appears to be the terminal default. Also, it's hard to figure out what
all of the names correlate in the theme.