Adding current configurations!
This are the things that I currently have on the PC I'm starting with
that I would like to propogate out to my other linux boxes.

I am going to have to change my tmux config I believe because I will
probably move to tmuxp and I want to also change the colors and the date
format to yyyy-mm-dd on the right.