Updated lvim config to give formatter more time.
Also had an issue where setting some of the object values for the
underlying lvim stuff will break the config. Had to do some detective
work to get that to work right. It turns out it was my setting for
having the git info in the nerdtree or gutter? Either way, it's fixed.