Updated python defaults to stop complaining about line lengths < 120.
Getting a lot of unnecessary red in my files when I'm doing one-off
projects. Wanted to have a better default.

I will need to double-check that this is not going to cause any problems
if I were to have different settings for specific projects where I have
flake8 loaded in a virtual environment though.