Updated astrovim configuration.
Trying to get to a good place with the configurations that also includes
the stuff that I had been relying on in lunarvim.

Running into issues with ruff not starting a server and giving feedback.
Adding pyright made ruff work all of a sudden.

But ruff wasn't compiling on my phone via termux, so I might get rid of
it.